requirements.txt
absl-py==2.1.0
accelerate==0.28.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.1.0
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1695989787169/work
cachetools==5.5.0
causal-conv1d==1.2.0.post2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1696001773319/work
chardet==5.2.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
DataProperty==1.0.1
datasets==2.18.0
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.9
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.7.0
eval_type_backport==0.2.0
evaluate==0.4.1
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.1
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.37.0
google-auth-oauthlib==0.4.6
grpcio==1.68.1
h5py==3.11.0
huggingface-hub==0.21.4
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1701026962277/work
importlib_metadata==8.5.0
importlib_resources==6.4.0
iopath==0.1.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.3
joblib==1.3.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.41.1
# Editable install with no version control (lm_eval==0.4.1)
-e /raid5/liujiayu/mamba-main/3rdparty/lm-evaluation-harness
lxml==5.1.0
mamba-ssm==1.2.0.post1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mbstrdecoder==1.1.3
mistune==3.1.1
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
multiprocess==0.70.16
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.1
ninja==1.11.1.1
nltk==3.8.1
numba==0.58.1
numexpr==2.8.6
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1687808301083/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pathvalidate==3.2.0
peft==0.9.0
pexpect==4.9.0
pickleshare==0.7.5
pillow @ file:///croot/pillow_1707233021655/work
pipenv==2024.4.1
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pooch==1.8.2
portalocker==2.8.2
prompt_toolkit==3.0.50
protobuf==5.26.0
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==15.0.1
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.11.1
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.19.1
pyparsing==3.1.2
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
pytablewriter==1.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.2.1
referencing==0.35.1
regex==2023.12.25
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
requests-oauthlib==2.0.0
responses==0.18.0
rouge_score==0.1.2
rpds-py==0.20.1
rsa==4.9
sacrebleu==2.4.1
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentry-sdk==2.19.2
setproctitle==1.3.4
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.3.7
sqlitedict==2.1.0
stack-data==0.6.3
sympy==1.12
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.4
tensorboard==2.12.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
timm==0.9.16
tinycss2==1.2.1
tokenizers==0.15.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.4.2
tqdm==4.66.2
tqdm-multiprocess==0.0.11
traitlets==5.14.3
transformers==4.38.2
triton==2.2.0
typepy==1.3.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1708239446578/work
virtualenv==20.29.2
wandb==0.19.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.6
word2number==1.1
xxhash==3.4.1
yacs==0.1.8
yarg==0.1.9
yarl==1.9.4
zipp==3.20.2
zstandard==0.22.0
